Referall documentation and refactoring
